# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Image Provider",
    "version": "staging",
    "description": "API for uploading and serving images for Retina API"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api"
    }
  ],
  "paths": {
    "/v1/images": {
      "get": {
        "summary": "List uploaded images",
        "description": "Retrieve images index listing from the server.",
        "responses": {
          "200": {
            "description": "Image listing retrieved successfully."
          },
          "404": {
            "description": "Image listing directory not found."
          }
        }
      }
    },
    "/v1/images/{imageName}": {
      "get": {
        "summary": "Get a static image by name",
        "description": "Retrieve a specific static image from the server by its name.",
        "parameters": [
          {
            "in": "path",
            "name": "imageName",
            "required": true,
            "description": "The name of the image file to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Static image retrieved successfully.",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Static image not found."
          }
        }
      }
    },
    "/v1/upload": {
      "post": {
        "summary": "Upload images",
        "description": "Upload images to the server.",
        "parameters": [
          {
            "name": "data_private",
            "in": "query",
            "required": false,
            "description": "Specifies if the data should be private.",
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "The file(s) to upload."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "context": {
                      "type": "object",
                      "properties": {
                        "imageUrls": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "URLs of the uploaded images."
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "A message describing the result of the upload."
                    },
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 200
                    }
                  }
                },
                "example": {
                  "context": {
                    "imageUrls": [
                      "http://localhost:8000/api/v1/images/<imgId>.<imgExtension>"
                    ]
                  },
                  "message": "File Uploaded Successfully",
                  "code": 200
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "413": {
            "description": "Payload too large. The uploaded file exceeds the specified limit."
          },
          "415": {
            "description": "Unsupported Media Type. The uploaded file format is not supported."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/v1/upload/delete": {
      "delete": {
        "summary": "Remove uploaded images",
        "description": "Remove files from filesystem which were uploaded.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "URLs of images to remove."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provided URLs file representation has been deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "context": {
                      "type": "object",
                      "properties": {
                        "removedImageUrls": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "URLs of the remove images."
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "Provided URLs has been deleted."
                    },
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 200
                    }
                  }
                },
                "example": {
                  "context": {
                    "removedImageUrls": [
                      "http://localhost:8000/api/v1/images/<imgId>.<imgExtension>"
                    ]
                  },
                  "message": "Provided URLs has been deleted",
                  "code": 200
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "415": {
            "description": "Unsupported Media Type. The uploaded file format is not supported."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/v1/upload/delete/private": {
      "delete": {
        "summary": "Remove uploaded private images",
        "description": "Removes all images which has been uploaded with data_private=true query parameter.",
        "responses": {
          "200": {
            "description": "All private images has been deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "All private images has been deleted."
                    },
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 200
                    }
                  }
                },
                "example": {
                  "message": "All private images has been deleted",
                  "code": 200
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}